(PC-27791)[API] feat: Make `get_object` behavior a part of the abstraction
This behavior belong to the scope of the `LocalProvider` abstraction, and thus
allowing to be overloaded/customize by children implementations
