+++ api/src/pcapi/local_providers/chunk_manager.py
@@ -1,19 +1,8 @@
 import pcapi.core.offers.models as offers_models
 from pcapi.local_providers.providable_info import ProvidableInfo
 from pcapi.models import Model
-from pcapi.repository.providable_queries import get_existing_object
 from pcapi.repository.providable_queries import insert_chunk
 from pcapi.repository.providable_queries import update_chunk
-
-
-def get_existing_pc_obj(
-    providable_info: ProvidableInfo, chunk_to_insert: dict, chunk_to_update: dict
-) -> offers_models.Product | offers_models.Offer | offers_models.Stock | None:
-    object_in_current_chunk = get_object_from_current_chunks(providable_info, chunk_to_insert, chunk_to_update)
-    if object_in_current_chunk is None:
-        return get_existing_object(providable_info.type, providable_info.id_at_providers)
-
-    return object_in_current_chunk
 
 
 def get_object_from_current_chunks(
